Rename hooks file and made correct ext.
