+++ packages/rapid-cmi5/src/lib/design-tools/rapidcmi5_mdx/plugins/image/useImageStyle.ts
